backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = "^2.6.2"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
sqlalchemy = "^2.0.27"
pymysql = "^1.1.0"
pydantic-settings = "^2.2.1"
aiomysql = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
